InstructorDashboard

+++ Views/AllCourses.php
@@ -1,7 +1,9 @@
 <?php
 session_start();
 include "../config/database.php";
-$InstructorId = $_SESSION['user_id'];
+if(isset($_SESSION['user_id'])){
+    $InstructorId = $_SESSION['user_id'];
+}
 $sql = "SELECT * FROM Courses";
 $result = mysqli_query($conn, $sql);
 if ($result && mysqli_num_rows($result) > 0) {
@@ -31,14 +33,24 @@ if ($result && mysqli_num_rows($result) > 0) {
 <header>
         <nav class="navbar">
             <!-- Logo -->
-            <a href="../index.php" class="logo">E-Learners</a>
+            <a href="home.php" class="logo">E-Learners</a>
 
             <!-- Links for Desktop -->
             <ul class="nav-links">
-                <li><a href="AllCourses.php">All Courses</a></li>
-                <li><a href="student.php">Enrolled Courses</a></li>
-                <li><a href="#">Certificates</a></li>
+                <?php if(isset($_SESSION['role']) && $_SESSION['role'] === "Instructor"): ?>
+                    <li><a href="InstructorCourses.php">All Courses</a></li>
+                <?php endif?>
+                <?php if(isset($_SESSION['role']) && $_SESSION['role'] === "Student"): ?>
+                    <li><a href="AllCourses.php">All Courses</a></li>
+                <?php endif?>
+                <?php if(isset($_SESSION['role']) && $_SESSION['role'] === "Student"): ?>
+                    <li><a href="../Controllers/EnrollCourseController.php">Enrolled Courses</a></li>
+                <?php endif?>
+                <?php if(isset($_SESSION['role']) && $_SESSION['role'] === "Student"): ?>
+                    <li><a href="#">Certificates</a></li>
+                <?php endif?>
                 <li><a href="contact.php">Contact Us</a></li>                
+
 
                 <?php if(isset($_SESSION['role']) && $_SESSION['role'] === "Admin"):?>
                     <li class="nav-item dropdown">
@@ -75,34 +87,18 @@ if ($result && mysqli_num_rows($result) > 0) {
                 <li><a href="#">Enrolled Courses</a></li>
                 <li><a href="#">Certificates</a></li>
                 <li><a href="contact.php">Contact Us</a></li>
-                
-                <?php if(isset($_SESSION['role']) && $_SESSION['role'] === "Admin"):?>
-                    <li class="nav-item dropdown">
-                    <a class="dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                        Content Management
-                    </a>
-                    <ul class="dropdown-menu">
-                        <li><a class="dropdown-item" href="../Areas/Admin/Courses.php">Courses</a></li>
-                        <li><a class="dropdown-item" href="../Areas/Admin/Users.php">Users</a></li>
-                    </ul>
-                </li>
-                <?php endif?>
-                
                 <?php
                         if(isset($_SESSION['user_id'])):?>
                             <li><a href="#" class="login-btn"> <?php echo $_SESSION['username'] ?> </a></li>
-                            <li><a href="../Controllers/Logout.php" class="register-btn">Log Out</a></li>
+                            <li><a href="Controllers/Logout.php" class="register-btn">Log Out</a></li>
                         <?php else:?>
-                            <li><a href="login.php" class="login-btn">Login</a></li>
-                            <li><a href="register.php" class="register-btn">Register</a></li>
+                            <li><a href="Views/login.php" class="login-btn">Login</a></li>
+                            <li><a href="Views/register.php" class="register-btn">Register</a></li>
                         <?php endif
                 ?>
-
-
             </ul>
         </nav>
     </header>
-
     <div class="container">
         <h1 class="text-center">Our Courses</h1>
         <div class="row mb-5">
@@ -124,7 +120,7 @@ if ($result && mysqli_num_rows($result) > 0) {
                         <p class="card-text mb-1"><strong>Category:</strong> Programming</p>
                         <p class="card-text mb-3"><strong>Price:</strong> <?= htmlspecialchars($course['Price']) ?></p>
                         <a href="course.php?id=<?= urlencode($course['Id']) ?>" class="btn btn-primary">View Details</a></div>
-                        <?php if($_SESSION['role'] === "Admin"):?>
+                        <?php if(isset($_SESSION['role']) && $_SESSION['role'] === "Admin"):?>
                             <a href="updatecourse.php?id=<?= urlencode($course['Id']) ?>" class=" w-25 m-auto mb-1 btn btn-primary">Edit</a>
                             <a href="deletecourse.php?id=<?= urlencode($course['Id']) ?>" class=" w-25 m-auto mb-1 btn btn-primary">Delete</a>
                             <?php endif?>
